refactor(design): extract ProtocolAuthor to lib/services/design/enums/
Pulls the role enum out of ward_common.dart into its own module so future
design services can depend on the enum without dragging in the
ward-common trailer logic. ward_common re-exports it, so existing
consumers (e.g. _ward_common_input.dart) keep working with a single
import.
