+++ lib/services/design/design_ward_common/ward_common.dart
@@ -1,18 +1,7 @@
 import '../_utils.dart';
+import '../enums/protocol_author.dart';
 
-/// The role of the person filling out a ward protocol form.
-///
-/// `label` doubles as the dropdown display text and the output line prefix
-/// (e.g. `Fellow: <name> (Tel: <tel>)`), so changes here propagate to both
-/// the UI and the rendered protocol.
-enum ProtocolAuthor {
-  resident('Resident'),
-  fellow('Fellow'),
-  staff('Staff');
-
-  final String label;
-  const ProtocolAuthor(this.label);
-}
+export '../enums/protocol_author.dart';
 
 /// Shared "Check ภาพ" + author identification inputs collected on every
 /// Protocol Ward screen (CVS, Chest, Abdomen, MSK, …).
